+++ scratch.js
@@ -1,4 +1,16 @@
 var nlp = require('./src/index');
 // nlp.verbose('tagger');
 
-let doc = nlp('The competent drum work of Don Brewer?').debug();
+//(private|general|major)
+
+// console.log(nlp('the so-called ❛singer-songwriter❜').normalize().out('text'));
+let doc = nlp(`"seven"`);
+// console.log(doc.list[0].terms[0]);
+console.log(doc.values().toNumber().out('text'));
+
+// console.log(nlp('Director of the F.B.I').acronyms().addPeriods().out('text'));
+
+// nlp('about this ...').normalize().debug()
+// console.log(nlp('word ...').toLowerCase().out());
+
+// console.log(nlp('word ...').list[0].terms);